Move the CLI package up a directory
This gets rid of an unnecessary part of the package's path (I didn't
know you could structure package this way).
