+++ cli/hcv1a/hcv1a.go
@@ -1,7 +1,7 @@
 package hcv1a
 
 import (
-	cli "github.com/hivdb/nucamino/cli/cli"
+	cli "github.com/hivdb/nucamino/cli"
 	data "github.com/hivdb/nucamino/data"
 )
